members cannot create or update memberships

+++ app/controllers/memberships_controller.rb
@@ -1,9 +1,11 @@
 class MembershipsController < SecretsController
   before_action :set_project
+  before_action :assert_owner, except: [:index, :destroy]
   def index
     @project = Project.find(params[:project_id])
     @memberships = @project.memberships
     @membership = Membership.new
+    @current_user_is_owner = current_user.owns_project?(@project)
   end
 
   def new
@@ -35,8 +37,10 @@ class MembershipsController < SecretsController
   end
 
   def destroy
+
     project = Project.find(params[:project_id])
     membership = Membership.find(params[:id])
+    assert_owner unless membership.user_id == current_user.id
     membership.destroy
     if membership.user_id == current_user.id
       redirect_to projects_path, notice: "#{current_user.full_name} successfully removed"
@@ -55,6 +59,12 @@ private
       unless @project and @project.users.include? current_user
         redirect_to projects_path,
         alert: "You do not have access to that project"
+      end
+    end
+
+    def assert_owner
+      unless current_user.owns_project?(@project)
+        redirect_to project_path(@project), alert: "You do not have access"
       end
     end
 end
